pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-commit"
version = "0.1.0"
description = "AI-powered Git workflow assistant with task management integration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["git", "commit", "ai", "jira", "automation", "workflow", "task-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
    "GitPython>=3.1.0",
    "requests>=2.31.0"
]

[project.urls]
Homepage = "https://github.com/ertiz82/smart-commit"
Documentation = "https://github.com/ertiz82/smart-commit#readme"
Repository = "https://github.com/ertiz82/smart-commit.git"
Issues = "https://github.com/ertiz82/smart-commit/issues"

[project.scripts]
smart-commit = "smart_commit.cli:main"
sgc = "smart_commit.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["smart_commit*"]